Feature: implement ApplyInputFilter method for various resource types
