--- pkg/rorresources/rortypes/resource_input_filter_slackmessage.go
@@ -1,6 +1,0 @@
-package rortypes
-
-// (r *ResourceSlackMessage) ApplyInputFilter Applies the input filter to the resource
-func (r *ResourceSlackMessage) ApplyInputFilter(cr *CommonResource) error {
-	return nil
-}
