--- pkg/rorresources/rortypes/resource_input_filter_vulnerabilityreport.go
@@ -1,11 +1,0 @@
-package rortypes
-
-import "github.com/NorskHelsenett/ror/pkg/config/globalconfig"
-
-// (r *ResourceVulnerabilityReport) ApplyInputFilter Applies the input filter to the resource
-func (r *ResourceVulnerabilityReport) ApplyInputFilter(cr *CommonResource) error {
-	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
-		cr.RorMeta.Internal = true
-	}
-	return nil
-}
