+++ pkg/rorresources/rortypes/resourcedef_appproject.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s applicationProject struct used with ArgoCD
 type ResourceAppProject struct {
 	Spec ResourceAppProjectSpec `json:"spec"`
@@ -8,4 +12,12 @@ type ResourceAppProjectSpec struct {
 	Description  string                               `json:"description"`
 	SourceRepos  []string                             `json:"sourceRepos"`
 	Destinations []ResourceApplicationSpecDestination `json:"destinations"`
+}
+
+// (r *ResourceAppProject) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceAppProject) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalAppProjects[cr.Metadata.Name] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
