+++ pkg/rorresources/rortypes/resourcedef_daemonset.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s deployment struct
 type ResourceDaemonSet struct {
 	Status ResourceDaemonSetStatus `json:"status"`
@@ -12,4 +16,12 @@ type ResourceDaemonSetStatus struct {
 	UpdatedNumberScheduled int `json:"updatedNumberScheduled"`
 	DesiredNumberScheduled int `json:"desiredNumberScheduled"`
 	CurrentNumberScheduled int `json:"currentNumberScheduled"`
+}
+
+// (r *ResourceDaemonSet) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceDaemonSet) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
