+++ pkg/rorresources/rortypes/resourcedef_exposedsecretreport.go
@@ -1,9 +1,21 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s namepace struct
 type ResourceExposedSecretReport struct {
 	Report ResourceVulnerabilityReportReport `json:"report"`
 }
 type ResourceExposedSecretReportReport struct {
 	Summary AquaReportSummary `json:"summary"`
+}
+
+// (r *ResourceExposedSecretReport) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceExposedSecretReport) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
