+++ pkg/rorresources/rortypes/resourcedef_ingress.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s namepace struct
 type ResourceIngress struct {
 	Spec   ResourceIngressSpec   `json:"spec"`
@@ -65,4 +69,12 @@ type ResourceIngressStatusLoadBalancer struct {
 type ResourceIngressStatusLoadBalancerIngress struct {
 	Hostname string `json:"hostname"`
 	Ip       string `json:"ip"`
+}
+
+// (r *ResourceIngress) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceIngress) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
