+++ pkg/rorresources/rortypes/resourcedef_policyreport.go
@@ -1,5 +1,10 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+	"time"
+)
+
 // K8s PolicyReport struct generated by kyverno
 type ResourcePolicyReport struct {
 	Results      []ResourcePolicyReportResults `json:"results"`
@@ -29,4 +34,15 @@ type ResourcePolicyReportSummary struct {
 	Pass  int `json:"pass"`
 	Skip  int `json:"skip"`
 	Warn  int `json:"warn"`
+}
+
+// (r *ResourcePolicyReport) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourcePolicyReport) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+
+	cr.RorMeta.LastReported = time.Now().Local().String()
+
+	return nil
 }
