+++ pkg/rorresources/rortypes/resourcedef_rbacassessmentreport.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s namepace struct
 type ResourceRbacAssessmentReport struct {
 	Report ResourceVulnerabilityReportReport `json:"report"`
@@ -21,4 +25,12 @@ type AquaReportScanner struct {
 	Name    string `json:"name"`
 	Vendor  string `json:"vendor"`
 	Version string `json:"version"`
+}
+
+// (r *ResourceRbacAssessmentReport) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceRbacAssessmentReport) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
