+++ pkg/rorresources/rortypes/resourcedef_replicaset.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s namepace struct
 type ResourceReplicaSet struct {
 	Spec   ResourceReplicaSetSpec   `json:"spec"`
@@ -25,4 +29,12 @@ type ResourceReplicaSetStatus struct {
 	AvailableReplicas int `json:"availableReplicas"`
 	ReadyReplicas     int `json:"readyReplicas"`
 	Replicas          int `json:"replicas"`
+}
+
+// (r *ResourceReplicaSet) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceReplicaSet) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
