+++ pkg/rorresources/rortypes/resourcedef_vulnerabilityreport.go
@@ -1,5 +1,9 @@
 package rortypes
 
+import (
+	"github.com/NorskHelsenett/ror/pkg/config/globalconfig"
+)
+
 // K8s namepace struct
 type ResourceVulnerabilityReport struct {
 	Report ResourceVulnerabilityReportReport `json:"report"`
@@ -26,4 +30,12 @@ type ResourceVulnerabilityReportReportVulnerability struct {
 	Link             string  `json:"primaryLink"`
 	InstalledVersion string  `json:"installedVersion"`
 	FixedVersion     string  `json:"fixedVersion"`
+}
+
+// (r *ResourceVulnerabilityReport) ApplyInputFilter Applies the input filter to the resource
+func (r *ResourceVulnerabilityReport) ApplyInputFilter(cr *CommonResource) error {
+	if globalconfig.InternalNamespaces[cr.Metadata.Namespace] {
+		cr.RorMeta.Internal = true
+	}
+	return nil
 }
